Tolerate bad entity extraction.
* Add tests for tolerating bad extraction.
* Substantially refactor llm_filter so we don't have a giant function in docset.py
* Also fix a minor bug where the tokenized filter was re-looking up the text in the element
  rather than using the txt variable.
